requirements.txt
Django==1.11
django-summernote==0.8.11.4
Pillow==6.0.0
pytz==2019.1
sorl-thumbnail==12.5.0
